DIABLO-104, admin can approve (and optionally schedule) from /course page
