+++ diablo/models/approval.py
@@ -138,6 +138,10 @@ class Approval(db.Model):
         return cls.query.filter_by(approved_by_uid=approved_by_uid, section_id=section_id, term_id=term_id).first()
 
     @classmethod
+    def get_approvals(cls, section_id, term_id):
+        return cls.query.filter_by(section_id=section_id, term_id=term_id).all()
+
+    @classmethod
     def get_approvals_per_section_ids(cls, section_ids, term_id):
         criteria = and_(cls.section_id.in_(section_ids), cls.term_id == term_id)
         return cls.query.filter(criteria).order_by(cls.created_at).all()
