+++ requirements.txt
@@ -24,7 +24,7 @@ pytest==5.4.1
 pytest-flask==1.0.0
 pytz==2020.1
 responses==0.10.14
-tox==3.14.6
+tox==3.15.0
 
 # For Xena
 selenium==3.141.0
