+++ src/api/course.ts
@@ -4,13 +4,15 @@ import utils from '@/api/api-utils'
 export function approve(
     publishType: string,
     recordingType: string,
-    sectionId: string
+    sectionId: string,
+    andSchedule?: boolean
 ) {
   return axios
     .post(`${utils.apiBaseUrl()}/api/course/approve`, {
       publishType,
       recordingType,
-      sectionId
+      sectionId,
+      andSchedule
     })
 }
