+++ src/views/Course.vue
@@ -89,7 +89,7 @@
                 Opted out
               </v-col>
             </v-row>
-            <v-row 
+            <v-row
               v-if="$currentUser.isAdmin && course.room && course.room.capability"
               id="send-invite"
               justify="center"
@@ -119,7 +119,7 @@
                     <v-card-text>The schedule and approvals for this course will be removed from Diablo, and the course will be marked as opt-out.</v-card-text>
                     <v-card-actions>
                       <v-spacer></v-spacer>
-                      <v-btn 
+                      <v-btn
                         id="confirm-unschedule-course-btn"
                         color="blue"
                         text
@@ -303,14 +303,24 @@
                   </div>
                 </v-col>
               </v-row>
-              <v-row v-if="!course.hasNecessaryApprovals" lg="2">
+              <v-row v-if="!course.hasNecessaryApprovals" lg="2" class="pr-6 pt-6">
                 <v-spacer />
+                <v-col v-if="$currentUser.isAdmin" md="4" class="mr-3">
+                  <v-btn
+                    id="btn-approve-and-schedule"
+                    color="outline"
+                    :disabled="disableSubmit"
+                    @click="approveAndSchedule"
+                  >
+                    Approve and Schedule
+                  </v-btn>
+                </v-col>
                 <v-col md="2">
                   <v-btn
                     id="btn-approve"
                     color="success"
                     :disabled="disableSubmit"
-                    @click="approveRecording"
+                    @click="approve"
                   >
                     Approve
                   </v-btn>
@@ -366,8 +376,13 @@
       }).catch(this.$ready)
     },
     methods: {
-      approveRecording() {
+      approve() {
         approve(this.publishType, this.recordingType, this.course.sectionId).then(data => {
+          this.render(data)
+        }).catch(this.$ready)
+      },
+      approveAndSchedule() {
+        approve(this.publishType, this.recordingType, this.course.sectionId, true).then(data => {
           this.render(data)
         }).catch(this.$ready)
       },
